+++ ruitk/tileWin/tileWin.js
@@ -31,12 +31,12 @@ export class TileWin {
         this.configStore = {
             tileSnapFirst : this.config.tileDirection,
             tileNudgeFirst : this.config.tileDirection,
-            tileRowDirection : this.config.tileDirection,
+            tileOppositeDirection : this.config.tileDirection,
             tileNudgeSwap : true,
-            maxX : this.config.tilePercentageX.length,
-            maxY : this.config.tilePercentageY.length,
+            xMax : this.config.tilePercentageX.length,
+            yMax : this.config.tilePercentageY.length,
         }
-        this.configStore.tileRowDirection = this.config.tileDirection === "y" ? "x" : "y";
+        this.configStore.tileOppositeDirection = this.config.tileDirection === "y" ? "x" : "y";
 
         return;
     }
@@ -87,39 +87,46 @@ export class TileWin {
     }
 
     update() {
-        let scrollRowMangerStyles = {
+        let rowManagerStyles = {
             transition : this.config.transition,
             backgroundColor : "rgba(0, 0, 0, 0)",
             borderColor : "rgba(0, 0, 0, 0)",
             position : "absolute",
+        };
+        let subRowManagerStyles = {
+            backgroundColor : "rgba(0, 0, 0, 0)",
+            borderColor : "rgba(0, 0, 0, 0)",
+            position : "relative",
+            display : "flex",
+            alignItems : "stretch",
         }
+        
+        function animateTile(id, x, y, w, h, t) {
+            if (t.config.animateOnCreateTile === true) {
+                Tile.transform(`tile${id}`, `calc(${x} + (${w} / 2))`, `calc(${y} + (${h} / 2))`, 0, 0);
+                setTimeout(() => {
+                    Tile.transform(`tile${id}`, x, y, w, h);
+                }, 0);
+            }
+        } 
         function makeTile(tile, id, x, y, w, h, p, content, t) {
             let { wInner, hInner } = calcWInnerAndHInner(tile, w, h, t);
-
-            if (t.config.animateOnCreateTile === true) {
-                if (t.config.tileRowType[tile[`${t.configStore.tileRowDirection}Snap`]] === "fixed") {
-                    Tile.create(`tile${id}`, `calc(${x} + (${wInner} / 2))`, `calc(${y} + (${hInner} / 2))`, 0, 0, t.fixedTileStyle, p);
-                } else {
-                    Tile.create(`tile${id}`, `calc(${x} + (${wInner} / 2))`, `calc(${y} + (${hInner} / 2))`, 0, 0, t.scrollTileStyle, p);
-                }
-                setTimeout(() => {
-                    Tile.transform(`tile${id}`, x, y, wInner, hInner);
-                }, 0);
+            
+            if (t.config.tileRowType[tile[`${t.configStore.tileOppositeDirection}Snap`]] === "fixed") {
+                Tile.create(`tile${id}`, x, y, wInner, hInner, t.fixedTileStyle, p);
             } else {
-                if (t.config.tileRowType[tile[`${t.configStore.tileRowDirection}Snap`]] === "fixed") {
-                    Tile.create(`tile${id}`, x, y, wInner, hInner, t.fixedTileStyle, p);
-                } else {
-                    Tile.create(`tile${id}`, x, y, wInner, hInner, t.scrollTileStyle, p);
-                }
+                Tile.create(`tile${id}`, x, y, wInner, hInner, t.scrollTileStyle, p);
             }
+            animateTile(id, x, y, wInner, hInner, t);
             if (tile.content !== null) {
-                Tile.append(`tile${tile.id}`, tile.content);
+                Tile.append(`tile${id}`, tile.content);
             }
         }
         function calcWInnerAndHInner(tile, w, h, t) {
             let wInner = `calc(${w} - (${Style.query("marginLeft", t.fixedTileStyle)} + ${Style.query("marginRight", t.fixedTileStyle)}))`;
             let hInner = `calc(${h} - (${Style.query("marginTop", t.fixedTileStyle)} + ${Style.query("marginBottom", t.fixedTileStyle)}))`;
-            if (t.config.tileRowType[tile[`${t.configStore.tileRowDirection}Snap`]] !== "fixed") {
+
+            if (t.config.tileRowType[tile[`${t.configStore.tileOppositeDirection}Snap`]] !== "fixed") {
                 if(t.config.tileDirection === "y") {
                     hInner = "auto";
                 } else {
@@ -131,8 +138,8 @@ export class TileWin {
         }
         
 
-        let tileLayout = List2D.create(this.configStore.maxX, this.configStore.maxY, false);
-        let tileLayoutLength = List2D.create(this.configStore.maxX, this.configStore.maxY, 0);
+        let tileLayout = List2D.create(this.configStore.xMax, this.configStore.yMax, false);
+        let tileLayoutLength = List2D.create(this.configStore.xMax, this.configStore.yMax, 0);
         // add items to layout
         for (let tile of this.tiles) {
             tileLayout[tile.xSnap][tile.ySnap] = true;
@@ -194,7 +201,7 @@ export class TileWin {
             }
 
             // Snap Minor axis
-            let xList = Array(this.configStore.maxX).fill(false);
+            let xList = Array(this.configStore.xMax).fill(false);
             for (let x in tileLayout) {
                 if(List2D.getListY(x, tileLayout).some(item => item === true)) {
                     xList[x] = true;
@@ -218,7 +225,7 @@ export class TileWin {
             }
 
             // Snap Minor axis
-            let yList = Array(this.configStore.maxY).fill(false);
+            let yList = Array(this.configStore.yMax).fill(false);
             for (let y in tileLayout[0]) {
                 if(List2D.getListX(y, tileLayout).some(item => item === true)) {
                     yList[y] = true;
@@ -312,7 +319,7 @@ export class TileWin {
         }
 
         /**
-         * The tilePercentageX & Y values and tileDistanceX & Y[-1] = 100 are necessary for calculating wSnap and hSnap.
+         * The tilePercentageX & Y values and tileDistanceX & Y[-1] = 100 are necessary for calculating wSnapPercent and hSnapPercent.
          * These calculations work by determining the x or y value at the end of the tile and subtracting the actual x or y value, leaving the difference.
          * Since there is technically no border between tiles, when calculating the end x and y values, it will reference the next row or column.
          * However, at the last row or column, you can't reference the next one, which would cause an error.
@@ -324,68 +331,139 @@ export class TileWin {
         tilePercentageX.push(0);
         let tilePercentageY = [...this.config.tilePercentageY];
         tilePercentageY.push(0);
-
-        // need better fix than this for scroll tile ordering as this only works when all scroll tiles are all rendered in the same update func call
-        this.tiles = this.tiles.sort((a, b) => {
-            if (a[`${this.config.tileDirection}Snap`] === b[`${this.config.tileDirection}Snap`]) {
-                return a[`${this.config.tileDirection}Nudge`] - b[`${this.config.tileDirection}Nudge`];
-            }
-            return a[`${this.config.tileDirection}Snap`] - b[`${this.config.tileDirection}Snap`];
-        });
         
-
+        
         // Make tiles
         for (let i in this.tiles) {
             let tile = this.tiles[i];
 
-            let xSnap = tileDistanceX[Math.floor(snapResize[tile.xSnap][tile.ySnap][0][0])]+
-                (tilePercentageX[Math.floor(snapResize[tile.xSnap][tile.ySnap][0][0])]*(snapResize[tile.xSnap][tile.ySnap][0][0]%1));
-            let ySnap = tileDistanceY[Math.floor(snapResize[tile.xSnap][tile.ySnap][0][1])]+
-                (tilePercentageY[Math.floor(snapResize[tile.xSnap][tile.ySnap][0][1])]*(snapResize[tile.xSnap][tile.ySnap][0][1]%1));
+            tile.xSnapPercent = tileDistanceX[Math.floor(snapResize[tile.xSnap][tile.ySnap][0][0])] +
+            (tilePercentageX[Math.floor(snapResize[tile.xSnap][tile.ySnap][0][0])] * (snapResize[tile.xSnap][tile.ySnap][0][0] % 1));
+            tile.ySnapPercent = tileDistanceY[Math.floor(snapResize[tile.xSnap][tile.ySnap][0][1])] +
+            (tilePercentageY[Math.floor(snapResize[tile.xSnap][tile.ySnap][0][1])] * (snapResize[tile.xSnap][tile.ySnap][0][1] % 1));
 
-            let wSnap = (tileDistanceX[Math.floor(snapResize[tile.xSnap][tile.ySnap][1][0])]+
-            (tilePercentageX[Math.floor(snapResize[tile.xSnap][tile.ySnap][1][0])]*(snapResize[tile.xSnap][tile.ySnap][1][0]%1))) - xSnap;
-            let hSnap = (tileDistanceX[Math.floor(snapResize[tile.xSnap][tile.ySnap][1][1])]+
-            (tilePercentageX[Math.floor(snapResize[tile.xSnap][tile.ySnap][1][1])]*(snapResize[tile.xSnap][tile.ySnap][1][1]%1))) - ySnap;
+            tile.wSnapPercent = (tileDistanceX[Math.floor(snapResize[tile.xSnap][tile.ySnap][1][0])] +
+            (tilePercentageX[Math.floor(snapResize[tile.xSnap][tile.ySnap][1][0])] * (snapResize[tile.xSnap][tile.ySnap][1][0] % 1))) - tile.xSnapPercent;
+            tile.hSnapPercent = (tileDistanceX[Math.floor(snapResize[tile.xSnap][tile.ySnap][1][1])] +
+            (tilePercentageX[Math.floor(snapResize[tile.xSnap][tile.ySnap][1][1])] * (snapResize[tile.xSnap][tile.ySnap][1][1] % 1))) - tile.ySnapPercent;
 
-            tile.x = xSnap + (wSnap * tile.snapShare[0][0]); // snapShare is for applying nudge to tiles
-            tile.y = ySnap + (hSnap * tile.snapShare[0][1]); // snapShare is relative value therefore the Snap vars just need to state how big the snap should be
+            tile.x = tile.xSnapPercent + (tile.wSnapPercent * tile.snapShare[0][0]); // snapShare is for applying nudge to tiles
+            tile.y = tile.ySnapPercent + (tile.hSnapPercent * tile.snapShare[0][1]); // snapShare is relative value therefore the Snap vars just need to state how big the snap should be
 
-            tile.w = wSnap * (tile.snapShare[1][0] - tile.snapShare[0][0]);
-            tile.h = hSnap * (tile.snapShare[1][1] - tile.snapShare[0][1]);
-            if (tile.status === "unrendered") {
-                if (this.config.tileRowType[tile[`${this.configStore.tileRowDirection}Snap`]] === "fixed") {
-                    makeTile(tile, tile.id, `${tile.x}%`, `${tile.y}%`, `${tile.w}%`, `${tile.h}%`, this.config.parent, tile.content, this);
-                } else {
-                    // create row manager
-                    if (document.querySelector(`#ScrollRowManger${tile[`${this.configStore.tileRowDirection}Snap`]}`) === undefined || document.querySelector(`#ScrollRowManger${tile[`${this.configStore.tileRowDirection}Snap`]}`) === null) {
-                        if (this.configStore.tileRowDirection === "x") {
-                            Tile.create(`ScrollRowManger${tile[`${this.configStore.tileRowDirection}Snap`]}`, `${xSnap}%`, 0, `${wSnap}%`, "auto", scrollRowMangerStyles, this.config.parent);
-                        } else {
-                            Tile.create(`ScrollRowManger${tile[`${this.configStore.tileRowDirection}Snap`]}`, 0, `${ySnap}%`, "auto", `${hSnap}%`, scrollRowMangerStyles, this.config.parent);
-                        }
-                    }
-                    if (this.configStore.tileRowDirection === "x") {
-                        makeTile(tile, tile.id, "0%", "0%", "100%", "auto", `#ScrollRowManger${tile[`${this.configStore.tileRowDirection}Snap`]}`, tile.content, this);
-                    } else {
-                        makeTile(tile, tile.id, "0%", "0%", "auto", "100%", `#ScrollRowManger${tile[`${this.configStore.tileRowDirection}Snap`]}`, tile.content, this);
-                    }
+            tile.w = tile.wSnapPercent * (tile.snapShare[1][0] - tile.snapShare[0][0]);
+            tile.h = tile.hSnapPercent * (tile.snapShare[1][1] - tile.snapShare[0][1]);
+        }
+
+        let sortedTiles = Array(this.configStore[`${this.configStore.tileOppositeDirection}Max`]).fill([]);
+        sortedTiles = JSON.parse(JSON.stringify(sortedTiles));
+        let somewhatSortedtiles = Array(this.configStore[`${this.configStore.tileOppositeDirection}Max`]).fill([]);
+        somewhatSortedtiles = JSON.parse(JSON.stringify(somewhatSortedtiles));
+        for (let tile of this.tiles) {
+            somewhatSortedtiles[tile[`${this.configStore.tileOppositeDirection}Snap`]].push(tile);
+        }
+        for (let i in somewhatSortedtiles) {
+            somewhatSortedtiles[i].sort((a , b) => {return a[`${this.config.tileDirection}Snap`] - b[`${this.config.tileDirection}Snap`]});
+            if (this.config.tileRowType[i] === "fixed") {
+                sortedTiles[i] = somewhatSortedtiles[i];
+                continue;
+            }
+            
+            let groupedByA = {}; // a = this.config.tileDirection
+            somewhatSortedtiles[i].forEach(tile => {
+                if (!groupedByA[tile[this.config.tileDirection]]) {
+                    groupedByA[tile[this.config.tileDirection]] = [];
                 }
-                tile.status = "rendered";
+                groupedByA[tile[this.config.tileDirection]].push(tile);
+            });
+
+            for (let a in groupedByA) {
+                groupedByA[a].sort((a, b) => a[this.configStore.tileOppositeDirection] - b[this.configStore.tileOppositeDirection]);
+            }
+
+            sortedTiles[i] = Object.values(groupedByA);
+        }
+
+        function tileUpdates(tile, t, i=0, j=0, k=0,) {
+            if (t.config.tileRowType[tile[`${t.configStore.tileOppositeDirection}Snap`]] === "fixed") {
+
+                let { wInner, hInner } = calcWInnerAndHInner(tile, `${tile.w}%`, `${tile.h}%`, t);
+                
+                Tile.transform(`tile${tile.id}`, `${tile.x}%`, `${tile.y}%`, wInner, hInner);
             } else {
-                if (this.config.tileRowType[tile[`${this.configStore.tileRowDirection}Snap`]] === "fixed") {
+                let tileElement = document.getElementById(`tile${tile.id}`);
+                let parentElementId = tileElement.parentElement.id;
+                
+                let match = parentElementId.match(/subRowManager(\d+)-(\d+)/);
+                let iOld = parseInt(match[1]);
+                let jOld = parseInt(match[2]);
 
-                    let { wInner, hInner } = calcWInnerAndHInner(tile, `${tile.w}%`, `${tile.h}%`, this);
-                    
-                    Tile.transform(`tile${tile.id}`, `${tile.x}%`, `${tile.y}%`, wInner, hInner);
+                tileElement.style.order = k;
+
+                if (iOld !== i || jOld !== j) {
+                    document.getElementById(`subRowManager${i}-${j}`).appendChild(tileElement);
                 }
-                if (tile.status === "contentChanged") {
-                    Tile.remove(`tile${tile.id}`);
-                    Tile.append(`tile${tile.id}`, tile.content);
+                
+            }
+
+            if (tile.status === "contentChanged") {
+                Tile.remove(`tile${tile.id}`);
+                Tile.append(`tile${tile.id}`, tile.content);
+            }
+        }
+
+        for (let i = 0; i < sortedTiles.length; i++) {
+            let row = sortedTiles[i];
+            
+            if (this.config.tileRowType[i] === "fixed") {
+                for (let j = 0; j < row.length; j++) {
+                    let tile = row[j];
+                    if (tile.status === "unrendered") {
+                        makeTile(tile, tile.id, `${tile.x}%`, `${tile.y}%`, `${tile.w}%`, `${tile.h}%`, this.config.parent, tile.content, this);
+                        tile.status = "rendered";
+                    } else {
+                        tileUpdates(tile, this);
+                    }
+                }
+                continue;
+            }
+
+            // create row manager
+            if (document.querySelector(`#rowManager${i}`) === undefined || document.querySelector(`#rowManager${i}`) === null) {
+                if (this.configStore.tileOppositeDirection === "x") {
+                    Tile.create(`rowManager${i}`, `${row[0][0].xSnapPercent}%`, 0, `${row[0][0].wSnapPercent}%`, "auto", rowManagerStyles, this.config.parent);
+                } else {
+                    Tile.create(`rowManager${i}`, 0, `${row[0][0].ySnapPercent}%`, "auto", `${row[0][0].hSnapPercent}%`, rowManagerStyles, this.config.parent);
+                }
+            }
+            for (let j = 0; j < row.length; j++) {
+                let subRow = row[j];
+                // create sub row manager
+                if (document.querySelector(`#subRowManager${i}-${j}`) === undefined || document.querySelector(`#subRowManager${i}-${j}`) === null) {
+                    if (this.configStore.tileOppositeDirection === "x") {
+                        Tile.create(`subRowManager${i}-${j}`, 0, 0, "100%", "auto", subRowManagerStyles, `#rowManager${i}`);
+                    } else {
+                        Tile.create(`subRowManager${i}-${j}`, 0, 0, "auto", "100%", subRowManagerStyles, `#rowManager${i}`);
+                    }
+                }
+                for (let k = 0; k < subRow.length; k++) {
+                    let tile = subRow[k];
+                    if (tile.status === "unrendered") {
+                        if (this.configStore.tileOppositeDirection === "x") {
+                            makeTile(tile, tile.id, "0%", "0%", `${(tile.snapShare[1][0] - tile.snapShare[0][0])*100}%`, "auto", `#subRowManager${i}-${j}`, tile.content, this);
+                        } else {
+                            makeTile(tile, tile.id, "0%", "0%", "auto", `${(tile.snapShare[1][0] - tile.snapShare[0][0])*100}%`, `#subRowManager${i}-${j}`, tile.content, this);
+                        }
+                        let tileElement = document.getElementById(`tile${tile.id}`);
+                        tileElement.style.order = k;
+                        tile.status = "rendered";
+                    } else {
+                        tileUpdates(tile, this, i, j, k);
+                    }
                 }
             }
         }
     }
+
 
     append(name, content) {
         if (content === null) {