+++ doc/changelog.md
@@ -1,3 +1,15 @@
+# 0.5
+#### style.js changes
+ - Added a new function called `declare` - [Documentation Here](style.js.md#Declare)
+#### documentation Changes 
+ - Added documentation changes for `query` and `declare` from `style.js`
+#### tile.js Changes
+ - Added a new function called `destory` to delete a tile
+#### tileWin.js Changes
+ - Scroll tiles will now take nudge into account when positioning thanks to a rewrite of how tile rendering and state handling work
+ - Renamed `tileRowDirection` to `tileOppositeDirection` in `configStore`
+ - Renamed `maxX` to `xMax` in `configStore`
+ - Renamed `maxY` to `yMax` in `configStore`
 ## 0.4.1
 #### tileWin.js Changes
  - Fixed a issue with scroll elements overlapping when there are multiple in the same row